+++ CHANGELOG.md
@@ -3,13 +3,12 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 ### Added
-
 - Homepage Hero component
 - Bees Tiers component
 - Hard Truth component
 - Open Source FAQ component
 - Solving Problems component
 - What We Stand For component
-
+- ltr, rtl tailwind animations
 ### Changed
 - Split index into components